:sparkles: Navigation on the large devices
Improve styles for navigation on the large devices. Remove unused
styles. Fix bug with fetching orders in the Orders view. Fix bug
with map coordinates validation on the admin view.

+++ client/src/app/views/Webpage/product/product.component.ts
@@ -90,14 +90,17 @@ export class ProductComponent implements OnInit, OnDestroy {
 
     try {
       const product: Product = await this.productsService.fetchProduct(this.id);
-      const { products } = await this.productsService.fetchProductsInCart(
-        this.viewedProducts,
-      );
 
-      if (products.length) {
-        this.viewedProductsService.setProducts(products);
-      } else {
-        this.viewedProductsService.removeProducts();
+      if (this.viewedProducts.length) {
+        const { products } = await this.productsService.fetchProductsInCart(
+          this.viewedProducts,
+        );
+
+        if (products.length) {
+          this.viewedProductsService.setProducts(products);
+        } else {
+          this.viewedProductsService.removeProducts();
+        }
       }
 
       this.product = {
